% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputFunctions.R
\name{build.PandF}
\alias{build.PandF}
\title{Construct time varying fecundity rates and birth ogives}
\usage{
build.PandF(
  Fdat = fecundity,
  Fproj = Fproj,
  Pdat = Pdat,
  Pper = Pper,
  years = years,
  return.periods = T
)
}
\arguments{
\item{Fdat}{Observed fecundity rates.}

\item{Fproj}{Which fecundity rate used in projections. Default is mean value over all observed fecundity rates ("mean"). If not a specified fecundity rate can be used (just insert number).}

\item{Pdat}{Observed birth ogives.}

\item{Pper}{For which time periods the birth ogives are valid.}

\item{years}{Start and stop year for the model (without projections).}

\item{return.periods}{Whether to return data frame containing information 
about time periods of P and F sampling. NOTE! This is to allow plotting fecundity curves,
distinguishing between real sampled values and interpolated between sampling periods. 
The returned object, Pper, must be omitted from data that are input into model.}
}
\value{
Pmatrix Time varying birth ogives.
}
\description{
Prepare the fecundity vector and the birth ogive matrix used by the model. A linear
transition between years with missing data is used.
}
\examples{
FvecandPmat <- build.PandF(fecundity,catch_data,Fproj,Pdat,Pper)
}
\keyword{Fecundity,}
\keyword{birth}
\keyword{ogive}
