% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catch.options.R
\name{eq.quota.helper.N1}
\alias{eq.quota.helper.N1}
\title{Internal helper function for finding equilibrium quota}
\usage{
eq.quota.helper.N1(Tot, dataN1, parametersN1, quota)
}
\arguments{
\item{Tot}{Total annual catch for population projections}

\item{quota}{Proportional catch of 0 and 1+ animals}

\item{dataD}{The data to be analyzed}

\item{parametersD}{Parameters used for the model}
}
\value{
minN Minimum found for difference between current and projected 1+ population size
}
\description{
Finding N-based equilibrium quota
}
\keyword{model}
\keyword{population}
