% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputFunctions.R
\name{load.data}
\alias{load.data}
\title{Load and prepare data}
\usage{
load.data(
  population = "harpeast",
  Amax = 20,
  years_of_prediction = 15,
  Fpast = "est",
  Fproj = "mean",
  catch_quota = c(0, 0),
  return_fec = F
)
}
\arguments{
\item{population}{Choose which population to run the model on (harpwest,harpeast,hooded).}

\item{Amax}{Maximum age group. Default is 20 years.}

\item{years_of_prediction}{Number of years in the future to project the model. Default is 15 years.}

\item{Fpast}{Which fecundity rate to use in past estimations. Fproj = "mean" uses mean value of observed fecundity rates. Otherwise available estimates (linearly interpolated) are used.}

\item{Fproj}{Which fecundity rate to use in future projections. Fproj = "mean" uses mean value of observed fecundity rates. Otherwise a fixed Fproj can be set.}

\item{catch_quota}{Catch quota for 0 and 1+animals to be used in future projections. catch_quota = "mean" uses mean value of observed catch for the last 5 years.}

\item{return_fec}{Return entire fecundity table, including information on years of sampling. Useful for plotting etc.}
}
\value{
data List of loaded data ready for TMB.
}
\description{
Load and prepare data to run the assessment model using TMB requirements.
}
\examples{
load.data(population = "harpeast")
}
\keyword{data}
\keyword{input,}
