% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputFunctions.R
\name{load.initial.values}
\alias{load.initial.values}
\title{Load initial values}
\usage{
load.initial.values(
  population = "harpeast",
  fromFile = TRUE,
  Kinit = 2e+06,
  Minit = 0.09,
  M0init = 0.27
)
}
\arguments{
\item{population}{Choose which population to run the model on (harpwest,harpeast,hooded).}

\item{fromFile}{Load initial values from file for specified population or set initial values manually.}

\item{Kinit}{Initial value for population size.}

\item{Minit}{Initial value for mortality of 1+ population, i.e., seals of age 1 or more.}

\item{M0init}{Initial value for pup mortality.}
}
\value{
parameters List of parameters with initial values to be estimated fromt the model.
}
\description{
Load initial values for which to start the optimization from. These are the values to be
estimated from the model.
}
\examples{
load.initial.values(population = "harpeast")
}
\keyword{initial}
\keyword{input,}
\keyword{parameters,}
\keyword{values}
