% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catch.options.R
\name{N70.helper.D}
\alias{N70.helper.D}
\title{Internal helper function for finding N70 quota}
\usage{
N70.helper.D(Tot, dataD, parametersD, quota, predYears)
}
\arguments{
\item{Tot}{Total annual catch for population projections}

\item{dataD}{The data to be analyzed}

\item{parametersD}{Parameters used for the model}

\item{quota}{Proportional catch of 0 and 1+ animals}

\item{predYears}{Define how many years the reduction will be carried out on. Default (10)}
}
\value{
minD07 Minimum found for difference between the depletion coefficient D and 0.7
}
\description{
Finding D-based N70 quota
}
\keyword{model}
\keyword{population}
