% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catch.options.R
\name{N70.helper.Nmax}
\alias{N70.helper.Nmax}
\title{Internal helper function for finding N70 quota}
\usage{
N70.helper.Nmax(Tot, dataNmax, parametersNmax, quota, predYears)
}
\arguments{
\item{Tot}{Total annual catch for population projections}

\item{quota}{Proportional catch of 0 and 1+ animals}

\item{predYears}{Define how many years the reduction will be carried out on. Default (10)}

\item{population}{Choose which population to run the model on (harpwest,harpeast,hooded).}
}
\value{
minN70 Minimum found for difference between N70 and the projected total population size
}
\description{
Finding N-based N70 quota
}
\keyword{model}
\keyword{population}
