% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catch.options.R
\name{PBR}
\alias{PBR}
\title{Function for finding PBR quota}
\usage{
PBR(
  n0 = n0,
  n1 = n1,
  se0 = se0,
  se1 = se1,
  rMax = 0.12,
  Fr = 0.5,
  quota = c(0.14, 1 - 0.14),
  cv = NA
)
}
\arguments{
\item{n0}{Estimated current population size of 0 animals}

\item{n1}{Estimated current population size of 1+ animals}

\item{se0}{Standard error of estimate for n0}

\item{se1}{Standard error of estimate for n1}

\item{rMax}{Maximum rate of population increase (by default 0.12, commonly used for pinnipeds)}

\item{Fr}{Assumed recovery factor}

\item{quota}{Proportional catch of 0 and 1+ animals}
}
\value{
Returns Minimum projected population size (Nmin) and Potential Biological Removal (PBR),
along with the PBR divided by 0 and 1+ animals given specified quota
}
\description{
Finding Potential Biological Removal (PBR) quotas
}
\keyword{model}
\keyword{population}
