% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputFunctions.R
\name{plotResCatch}
\alias{plotResCatch}
\title{Plot the model results and the reported catch data}
\usage{
plotResCatch(
  results = res,
  data = data,
  component = c("N0", "N1", "Tot"),
  xLim = NA,
  yLim = NA,
  plot.legend = TRUE,
  plot.Nlims = TRUE,
  projections = TRUE,
  mean.proj = FALSE,
  width = 13,
  height = 13,
  conf.int = TRUE,
  grDev = FALSE,
  labels = "English",
  long.labels = TRUE
)
}
\arguments{
\item{results}{Output from fitted model}

\item{data}{Original data on estimated pup production (set to NA if these are not to be included).}

\item{component}{Which population component to plot. Can be either 'N0' for pups, 'N1' for adults, or a vector with both.}

\item{xLim}{Manually set the x axis extent, overrides the default which is the extent of the plotted data.}

\item{yLim}{Manually set the y axis extent, overrides the default which is the extent of the plotted data.}

\item{plot.legend}{Add legend to plot (TRUE/FALSE)}

\item{plot.Nlims}{True/False}

\item{projections}{Plot projections (TRUE/FALSE)}

\item{mean.proj}{True/False}

\item{width}{Width of the figure}

\item{height}{Height of the figure}

\item{conf.int}{Logical parameter to decide wether to plot 95 percent confidence interval or not}

\item{grDev}{Logical parameter to decide wether to open a OS independent graphical window}

\item{long.labels}{Set TRUE for more detailed legend and FALSE for simple}
}
\value{
plot Returns a plot of predicted population size for different population components
}
\description{
Plot the model results and the reported catch data in the same figure.
}
\examples{
plotCResatch(data$Cdata)
}
\keyword{model}
\keyword{population}
